requirements.txt
amqp==2.4.1
anyjson==0.3.3
billiard==3.5.0.5
celery==4.2.1
certifi==2018.1.18
chardet==3.0.4
Django==2.1.2
django-celery==3.2.2
django-suit==0.2.26
djangorestframework==3.9.0
gevent==1.4.0
greenlet==0.4.15
idna==2.6
kombu==4.3.0
liqpay-python3==1.1.1
Pillow==5.0.0
pytz==2018.9
redis==3.1.0
requests==2.18.4
urllib3==1.22
vine==1.2.0
